requirements.txt
Django>=4.2
djangorestframework
pyswisseph
python-dateutil
gunicorn
whitenoise
pytz